qemu: Add QEMU 2.10 x86_64 the generated capabilities
For reference, these were generated by updating a local qemu git
repository to the latest upstream and building in order to generate
an "x86_64-softmmu/qemu-system-x86_64" image.

The image was then provided as input:

    tests/qemucapsprobe /path/to/x86_64-softmmu/qemu-system-x86_64 > \
       tests/qemucapabilitiesdata/caps_2.10.0.x86_64.replies

The image was then used in a test domain as the emulator and the
domain started using a "top of tree" build running libvirtd which
generates an XML file in /var/cache/libvirt/qemu/capabilities/.
I search the cache for 2.10:

    grep 2.10 /var/cache/libvirt/qemu/capabilities/*.xml

and copied the resulting xml file to

    tests/qemucapabilitiesdata/caps_2.10.0.x86_64.xml

scrubbed the file slightly to set qemuctime, selfctime, and
selfvers to 0 and change the package to add a "-dirty".

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170831110156.11549-2-jferlan@redhat.com>