storage: Check and possibly update config file after build
https://bugzilla.redhat.com/show_bug.cgi?id=1464313

As it turns out, the on-disk config file (e.g. not the stateDir file)
needs to be updated when --override is provided since it's possible and
highly probable that the def->source.format has been adjusted and could
cause a future start after perhaps a libvirtd restart to have the older
format from a define operation from the backend.

So in the 2 places where it's possible a write would be needed (create
after define and build), let's perform a check and do the save/write
operation on the configFile if it's necessary.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170906192418.32358-1-jferlan@redhat.com>