iohelper: fix reading with O_DIRECT
saferead is not suitable for direct reads. If file size is not multiple
of align size then we get EINVAL on the read(2) that is supposed to
return 0 because read buffer will not be aligned at this point.

Let's not read again after partial read and check that we read
everything by comparing the number of total bytes read against file size.
Message-Id: <1504770256-674340-5-git-send-email-nshirokovskiy@virtuozzo.com>