conf: Add USB companion controllers with virDomainControllerInsert()
virDomainDefAddController() is very convenient, but it is only passed
a limited number of information about the controller.

When adding USB controllers, knowing whether the controller is a
master or a companion is important because it can influence the order
devices show up on the QEMU command line, and ultimately whether the
guest can be started at all.

To make sure USB controllers will always be sorted correctly,
allocate the virDomainControllerDef structure explicitly, fill in the
relevant information manually and call virDomainControllerInsert()
directly.

Clean up both virDomainDefAddUSBController() itself and
virDomainDefAddController() while at it. The behavior is not altered,
as evidenced by the lack of updates to the test suite.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20170907170013.13474-2-abologna@redhat.com>