travis: Don't abort build due to -Wvariadic-macros
The openwsman header files are at fault here, but precise is entirely
unmaintained at this point so the issue will never be fixed.

Better to ignore the error and have coverage over the Hyper-V driver
than disabling it: if code that would trigger the warning will be
added to libvirt, the CentOS CI will catch it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20170914130908.13067-4-abologna@redhat.com>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>