conf: Add/Allow parsing the auth in the disk source
Since the virStorageAuthDefPtr auth; is a member of _virStorageSource
it really should be allowed to be a subelement of the disk <source>
for the RBD and iSCSI prototcols. That way we can set up to allow
the <auth> element to be formatted within the disk source.

For now just allow the format in the RNG and read it in domain_conf.

Modify the qemuxml2argvtest to add a parse failure when there is an
<auth> as a child of <disk> *and* an <auth> as a child of <source>.

The virschematest will read the new test files and validate from a
RNG viewpoint things are fine

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170916003017.32502-4-jferlan@redhat.com>