conf: Move auth formatting to disk source
Alter the output of the formatting to be a child of the disk's source
rather than a child of the disk.

Update the various test outputs for existing disk tests to conform to
the new view.

Add tests to validate that if the <auth> was found in <source>, then
the resulting xml2xml and xml2arg works just fine.  The two new .args
file are exact copies of the non "-source" version of the file.

Update the virstoragetest to handle that the <auth> output will now be
part of the <source> stanza in the rbd output.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170916003017.32502-5-jferlan@redhat.com>