conf: Move LUKS encryption formatting to disk source
Alter the output of the formatting to be a child of the disk's source
rather than a child of the disk for LUKS encryption, but keep the legacy
QCOW encryption as a child of disk.

Update the various test outputs for existing disk tests to conform to
the new view. The qemuxml2xmlout-luks-disks.xml used to be a link to
the "source" tests/qemuxml2argvdata/qemuxml2argv-luks-disks.xml, but
with the change to move LUKS output to a child of <source>, the output
file now will differ. While a link would still work, that would require
changing the source file which wasn't the goal.

Add tests to validate that if the <encryption> was found in <source>, then
the resulting xml2xml and xml2arg works just fine.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170916003017.32502-9-jferlan@redhat.com>