qemu: Add disk secret object hash table to _qemuDomainObjPrivate
Currently when an AES secret object is added to the domain for
either a network disk, a LUKS encryption secret, or for a SCSI
hostdev there is no way for domain restart to be able to connect
or determine which secret by secrettype and uuid or usage was
used in order to generate the object.

So, in order to be able to lookup which secret generated the
object, this patch will create and manage a private object hash
table that tracks when a disk using the secret object is added
or removed in order to be able to "rebuild" the secret.

This information will be recorded in the domain private XML file
so that when libvirtd restarts, we can rebuild and determine which
secret was used.

The qemuDomainObjDiskSecretObjectAliasEntryLookup helper is
currently unused, but it's purpose would be to find the object
alias in the table and return the usageType and seclookupdef
that created the object. Could be quite useful for something.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170916003017.32502-12-jferlan@redhat.com>