nodedev: Introduce udevEventCheckMonitorFD helper function
We need to perform some sanity checks on the udev monitor before every
use so that we know nothing changed in the meantime. The reason for
moving the code to a separate function is to be able to perform the same
check from a worker thread that will replace the udevEventHandleCallback
in terms of poking udev for device events.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <e89b30fa65948a8757189cb946dfbaca4c082aa7.1505752213.git.eskultet@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>