nodedev: Disable/re-enable polling on the udev fd
The event loop may get scheduled earlier than the udev event handler
thread which means that it would keep invoking the handler callback with
"new" events, while in fact it's most likely still the same event which
the handler thread hasn't managed to remove from the socket queue yet.
This is due to unwanted increments of the number of events queuing on the
socket and causes the handler thread to spam the logs with an error about
libudev returning NULL (errno == EAGAIN).

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <3d1af4740817c4ca8ca787c42f2ea4f96a99fc44.1505752213.git.eskultet@redhat.com>