util: Introduce virFileWaitForExists
Since we have a number of places where we workaround timing issues with
devices, attributes (files in general) not being available at the time
of processing them by calling usleep in a loop for a fixed number of
tries, we could as well have a utility function that would do that.
Therefore we won't have to duplicate this ugly workaround even more.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <70d602d2ff7489ccea9bbd6734a55e20db6be2b4.1505752213.git.eskultet@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>