qemu: block: Use correct alias when extracting disk node names
For some arcane reason we don't use the alias from disk->info.alias
directly but prepend drive in front of it. This messed up the disk node
name extraction code as qemu reports the full alias. This was broken in
the node name detector refactoring done in commit 0175dc6ea024d

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1494327
Message-Id: <ab2599d3fa68b8788264dc66d387f007b45ea794.1506333120.git.pkrempa@redhat.com>