qemu: process: Refresh data from qemu monitor after migration
Some values we read from the qemu monitor may be changed with the actual
state by the incomming migration. This means that we should refresh
certain things only after the migration has finished.

This is mostly visible in the cdrom tray state, which is by default
closed but may be opened by the guest OS. This would be refreshed before
qemu transferred the actual state and thus libvirt would think that the
tray is closed.

Note that this patch moves only a few obvious query commands. Other may
be moved later after individual assesment.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1463168
Message-Id: <ddf9aa091014a2a89a952b7847f7d514b43df18f.1506349510.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>