qemu: hotplug: Ignore cgroup errors when hot-unplugging vcpus
When the vcpu is successfully removed libvirt would remove the cgroup.
In cases when removal of the cgroup fails libvirt would report an error.

This does not make much sense, since the vcpu was removed and we can't
really do anything with the cgroup. This patch silences the errors from
cgroup removal.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1462092
Message-Id: <7ebdaa52abb13b24826f90c25ae43d96dc2eb9f3.1506371847.git.pkrempa@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>